refactor: moved all cards to summarycards component
